move infrastructure layer into twimi-app and adapt all modules
