+++ twimi-app/src/infrastructure/repositories/health_check.rs
@@ -3,7 +3,9 @@ use std::sync::Arc;
 use async_trait::async_trait;
 use shaku::Component;
 
-use crate::{domain::repositories::health_check::HealthCheckRepository, infrastructure::Database};
+use twimi_core::domain::repositories::health_check::HealthCheckRepository;
+
+use crate::infrastructure::Database;
 
 #[derive(Component)]
 #[shaku(interface = HealthCheckRepository)]
@@ -14,7 +16,7 @@ pub struct HealthCheckRepositoryImpl {
 
 #[async_trait]
 impl HealthCheckRepository for HealthCheckRepositoryImpl {
-    async fn health_check(&self) -> Result<(), sqlx::Error> {
+    async fn health_check(&self) -> Result<(), anyhow::Error> {
         sqlx::query("SELECT 1")
             .execute(self.database.pool())
             .await?;
