+++ twimi-app/src/routes/mod.rs
@@ -4,4 +4,4 @@ mod signup;
 pub use health_check::health_check;
 pub use signup::signup;
 
-pub type Inject<T> = shaku_actix::Inject<twimi_core::AppModule, T>;
+pub type Inject<T> = shaku_actix::Inject<crate::AppModule, T>;
