+++ twimi-core/src/domain/repositories/health_check.rs
@@ -3,5 +3,5 @@ use shaku::Interface;
 
 #[async_trait]
 pub trait HealthCheckRepository: Interface {
-    async fn health_check(&self) -> Result<(), sqlx::Error>;
+    async fn health_check(&self) -> Result<(), anyhow::Error>;
 }
