+++ twimi-core/src/domain/repositories/mod.rs
@@ -1,9 +1,2 @@
 pub mod health_check;
 pub mod users;
-
-use shaku::Interface;
-use sqlx::MySqlPool;
-
-pub trait Database: Interface {
-    fn pool(&self) -> &MySqlPool;
-}
